refactored fhir facade into services, deleted patient
